% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{arrow_plot}
\alias{arrow_plot}
\title{Arrow plot}
\usage{
arrow_plot(m, p = 0.05)
}
\arguments{
\item{m}{A \code{lm} object}

\item{p}{p-value cutoff, highest p-value of independent variables you'd like to keep in your plot. defaults to .05}
}
\description{
Plot a model with arrows.
In this plot, the arrows can represent both the direction and magnitude of the effects.
The shaded rectangles represent a standard confidence interval.
If the shaded interval crosses the dotted line, there is no significant evidence suggesting the independent variable has a nonzero effect.
}
