% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{as.data.frame.lm}
\alias{as.data.frame.lm}
\title{Method to convert a linear model object into a data-frame}
\usage{
\method{as.data.frame}{lm}(m)
}
\arguments{
\item{m}{A \code{lm} object}
}
\description{
Method to convert a linear model object into a data-frame
}
