% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.lm}
\alias{plot.lm}
\title{Method to convert a linear model object into a plot}
\usage{
\method{plot}{lm}(m, ...)
}
\arguments{
\item{m}{A \code{lm} object}
}
\description{
Method to convert a linear model object into a plot
}
