% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot a histogram of residuals}
\usage{
plot_residuals(m, outlier = FALSE)
}
\arguments{
\item{m}{A \code{lm} object}
}
\description{
Plot a histogram of residuals
}
