% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{residual_dots}
\alias{residual_dots}
\title{Residual dots}
\usage{
residual_dots(m)
}
\arguments{
\item{m}{A \code{lm} object}
}
\description{
Plot actual vs fitted values, with orange lines to represent residuals.
Warning: this plot will look funky if you have a binary dependent variable!
}
