% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stat_table}
\alias{stat_table}
\title{Function to make a standard regression output table}
\usage{
stat_table(m)
}
\arguments{
\item{m}{A \code{lm} object}
}
\description{
Function to make a standard regression output table
}
